feat: rename platform lib to platforms
