+++ bin/tldr
@@ -4,7 +4,7 @@ const program = require('commander');
 const pkg = require('../package');
 const Tldr = require('../lib/tldr');
 const config = require('../lib/config');
-const platforms = require('../lib/platform');
+const platforms = require('../lib/platforms');
 const { TldrError } = require('../lib/errors');
 
 program
