+++ lib/cache.js
@@ -4,7 +4,7 @@ const fs = require('fs-extra');
 const os = require('os');
 const path = require('path');
 const remote = require('./remote');
-const platforms = require('./platform');
+const platforms = require('./platforms');
 const index = require('./index');
 const utils = require('./utils');
