+++ lib/config.js
@@ -4,7 +4,7 @@ const defaults = require('lodash/defaults');
 const fs = require('fs');
 const path = require('path');
 const osHomedir = require('os').homedir;
-const platforms = require('./platform');
+const platforms = require('./platforms');
 
 exports.get = () => {
   const DEFAULT = path.join(__dirname, '..', 'config.json');
