+++ lib/search.js
@@ -7,7 +7,7 @@ const natural = require('natural');
 const config = require('./config');
 const utils = require('./utils');
 const index = require('./index');
-const platforms = require('./platform');
+const platforms = require('./platforms');
 
 const CACHE_FOLDER = path.join(config.get().cache, 'cache');
