+++ lib/tldr.js
@@ -7,7 +7,7 @@ const ora = require('ora');
 const { EmptyCacheError, MissingPageError, MissingRenderPathError } = require('./errors');
 const Cache = require('./cache');
 const search = require('./search');
-const platforms = require('./platform');
+const platforms = require('./platforms');
 const parser = require('./parser');
 const render = require('./render');
 const index = require('./index');
